pcf_scripts/backend/requirements.txt
# CEAPSI Backend Dependencies

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database and storage
httpx>=0.25.0

# Data processing (shared with main app)
pandas>=2.0.3
numpy>=1.24.3
scikit-learn>=1.3.0
prophet>=1.1.5
statsmodels>=0.14.0
openpyxl>=3.1.2

# Configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0